% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.apiInfo}
\alias{print.apiInfo}
\title{Print the result of a GitHub API call}
\usage{
\method{print}{apiInfo}(x, ...)
}
\arguments{
\item{x}{The result object from listRepos()}

\item{...}{Other arguments to print generic.}
}
\value{
The tibble result with repos and description columns.
}
\description{
Print the result of a GitHub API call
}
