% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenAuthentication.r
\name{tokenAuthentication}
\alias{tokenAuthentication}
\title{Token authentication}
\usage{
tokenAuthentication(userName = NULL, token = NULL)
}
\arguments{
\item{userName}{Must be entered as a single string without whitespace, the violation will throw a dedicated error.}

\item{token}{Must be entered as a single string without whitespace, the violation will throw a dedicated error.}
}
\value{
tokenAuthentication() for authenticated user return a list of credentials (user-name, token, url) along with httr authentication() and response() objects
}
\description{
This function checks if the given credentials are entered properly to enable the connection to the JSON file.
Submitting a non-existing user-name or token will result in an unauthorized access error.
Function intentionally forces the user to enter proper credentials.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Wrongly entered credintials
tokenAuthentication("Mickey mouse", ".")
\dontshow{\}) # examplesIf}
}
