% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReposDescription.r
\name{updateReposDescription}
\alias{updateReposDescription}
\title{Title}
\usage{
updateReposDescription(x, reposToUpdate, newDescription)
}
\arguments{
\item{newDescription}{}
}
\description{
Title
}
